+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 From version 1.2.0 the file IP2LOCATION-LITE-DB5.BIN is no longer part of the docker image and will need to be downloaded (after registration) from [ip2location](https://lite.ip2location.com/database/ip-country?lang=en_US) and be placed in /srv/incore_auth.
 
+# Unreleased
+
+## Added
+- information about user and groups is synced every 30 minutes back to the database and datawolf
+
+## Changed
+- updated all packages used
+
 # [1.5.0] - 2022-09-24
 
 ## Changed